refactor worker
 - add redis start to start() function
 - remove helper function to initiate class
 - log mongo errors
 - rename database conf for userdata
